fix(#133) : rename file `marker_generator_service`
