+++ lib/services/marker_generator_service.dart
